+++ kernel/kernel/system/ni.c
@@ -1,7 +1,7 @@
 #include <kstdio.h>
 #include <system.h>
 
-int sys_ni(unsigned nr)
+long sys_ni(unsigned nr)
 {
     kprintf("Syscall (%#x) not implemented\n", nr);
